Moved gps_conversion to common/ directory
